Dealing with imbalance dataset
To deal with imbalance dataset there are two techniques:

Under Sampling: In this form, data is randomly selected from the normal category to match the number of fraud data with a 50:50 per cent ratio
Oversampling: In this form, we add more data to the Fraud category, we add data in the same dimension. Suppose the ratio of normal vs fraud transaction is 9:1, so we would add 9 times the same data for fraud to make it even.
There is one library in python which is used to deal with imbalance dataset which is 'imblearn' and it has two functions that deal with different types of handling imbalance dataset technique: nearmiss(for under samping) and SMOTETomake(over sampling).